pyproject.toml
[tool.poetry]
name = "audios"
version = "0.1.0"
description = ""
authors = ["SAwckA <saw.cka@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {extras = ["voice"], version = "^2.3.2"}
yt-dlp = "^2024.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
